+++ src/test/kotlin/io/jawg/geojson/MultiPointTest.kt
@@ -1,7 +1,7 @@
 package io.jawg.geojson
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import org.junit.Test
+import tools.jackson.module.kotlin.jacksonObjectMapper
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue