+++ src/test/kotlin/io/jawg/geojson/PolygonTest.kt
@@ -1,10 +1,10 @@
 package io.jawg.geojson
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import io.jawg.geojson.utils.GeoJsonLoader
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.skyscreamer.jsonassert.JSONAssert
+import tools.jackson.module.kotlin.jacksonObjectMapper
 import kotlin.test.assertFailsWith
 
 class PolygonTest {