+++ src/test/kotlin/io/jawg/geojson/serializer/PositionSerializerTest.kt
@@ -1,8 +1,8 @@
 package io.jawg.geojson.serializer
 
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import io.jawg.geojson.Position
 import org.junit.Test
+import tools.jackson.module.kotlin.jacksonObjectMapper
 import kotlin.test.assertEquals
 
 class PositionSerializerTest {